+++ pkg/webui/console/components/uplink-form/connect.js
@@ -16,17 +16,24 @@ import { connect } from 'react-redux'
 
 import api from '@console/api'
 
-import { selectSelectedApplicationId } from '@console/store/selectors/applications'
-import { selectSelectedDeviceId } from '@console/store/selectors/devices'
+import {
+  selectSelectedApplicationId,
+  selectApplicationLinkSkipPayloadCrypto,
+} from '@console/store/selectors/applications'
+import { selectSelectedDeviceId, selectSelectedDevice } from '@console/store/selectors/devices'
 
 const mapStateToProps = state => {
   const appId = selectSelectedApplicationId(state)
   const devId = selectSelectedDeviceId(state)
+  const device = selectSelectedDevice(state)
+  const skipPayloadCrypto = selectApplicationLinkSkipPayloadCrypto(state)
 
   return {
     appId,
     devId,
+    device,
     simulateUplink: uplink => api.device.simulateUplink(appId, devId, uplink),
+    skipPayloadCrypto,
   }
 }