+++ pkg/webui/console/components/uplink-form/uplink-form.js
@@ -15,6 +15,7 @@
 import React from 'react'
 import { defineMessages } from 'react-intl'
 
+import Notification from '@ttn-lw/components/notification'
 import SubmitButton from '@ttn-lw/components/submit-button'
 import Input from '@ttn-lw/components/input'
 import SubmitBar from '@ttn-lw/components/submit-bar'
@@ -50,7 +51,7 @@ const validationSchema = Yup.object({
 const initialValues = { f_port: 1, frm_payload: '' }
 
 const UplinkForm = props => {
-  const { simulateUplink } = props
+  const { simulateUplink, device, skipPayloadCrypto } = props
 
   const [error, setError] = React.useState('')
 
@@ -81,14 +82,20 @@ const UplinkForm = props => {
     [simulateUplink],
   )
 
+  const deviceSimulationDisabled = device.skip_payload_crypto_override || skipPayloadCrypto
+
   return (
     <>
+      {deviceSimulationDisabled && (
+        <Notification content={sharedMessages.deviceSimulationDisabledWarning} warning small />
+      )}
       <IntlHelmet title={m.simulateUplink} />
       <Form
         error={error}
         initialValues={initialValues}
         validationSchema={validationSchema}
         onSubmit={handleSubmit}
+        disabled={deviceSimulationDisabled}
       >
         <Form.SubTitle title={m.simulateUplink} />
         <Form.Field
@@ -117,7 +124,13 @@ const UplinkForm = props => {
 }
 
 UplinkForm.propTypes = {
+  device: PropTypes.device.isRequired,
   simulateUplink: PropTypes.func.isRequired,
+  skipPayloadCrypto: PropTypes.bool,
+}
+
+UplinkForm.defaultProps = {
+  skipPayloadCrypto: false,
 }
 
 export default UplinkForm